Widgetize the KPI progress graph
